Revert "vvfat: introduce a no-mbr option and handle last block data"
This reverts commit 2de04ac276c7285c792177c045c7ca40bef9ba8a.

This commit gathers in fact two commits:
 - introduce a no-mbr option, which is useful and will be added back in
   a cleaner patch.
 - handle last block data, which seems to trigger some assertion
   failures.

qemu#55

+++ block/vvfat.c
@@ -273,7 +273,6 @@ typedef struct mapping_t {
          */
         struct {
             uint32_t offset;
-            char *last_cluster_data;
         } file;
         struct {
             int parent_mapping_index;
@@ -286,8 +285,9 @@ typedef struct mapping_t {
     enum {
         MODE_UNDEFINED = 0,
         MODE_NORMAL = 1,
-        MODE_DIRECTORY = 2,
-        MODE_DELETED = 4,
+        MODE_MODIFIED = 2,
+        MODE_DIRECTORY = 4,
+        MODE_DELETED = 8,
     } mode;
     int read_only;
 } mapping_t;
@@ -849,7 +849,6 @@ static int read_directory(BDRVVVFATState* s, int mapping_index)
             } else {
                 s->current_mapping->mode = MODE_UNDEFINED;
                 s->current_mapping->info.file.offset = 0;
-                s->current_mapping->info.file.last_cluster_data = NULL;
             }
             s->current_mapping->path=buffer;
             s->current_mapping->read_only =
@@ -891,11 +890,7 @@ static int read_directory(BDRVVVFATState* s, int mapping_index)
 
 static inline int32_t sector2cluster(BDRVVVFATState* s,off_t sector_num)
 {
-    if (sector_num < s->offset_to_root_dir) {
-        return -1;
-    } else {
-        return (sector_num-s->offset_to_root_dir) / s->sectors_per_cluster;
-    }
+    return (sector_num - s->offset_to_root_dir) / s->sectors_per_cluster;
 }
 
 static inline off_t cluster2sector(BDRVVVFATState* s, uint32_t cluster_num)
@@ -927,7 +922,6 @@ static int init_directories(BDRVVVFATState *s, const char *dirname,
     s->sectors_per_fat=(s->sector_count+i)/i; /* round up */
 
     s->offset_to_fat = s->offset_to_bootsector + 1;
-
     s->offset_to_root_dir = s->offset_to_fat + s->sectors_per_fat * 2;
 
     array_init(&(s->mapping),sizeof(mapping_t));
@@ -1094,11 +1088,6 @@ static QemuOptsList runtime_opts = {
             .help = "Make the image writable",
         },
         {
-            .name = "no-mbr",
-            .type = QEMU_OPT_BOOL,
-            .help = "Do not add MBR on this disk",
-        },
-        {
             .name = BLOCK_OPT_SIZE,
             .type = QEMU_OPT_SIZE,
             .help = "Virtual disk size"
@@ -1113,7 +1102,6 @@ static void vvfat_parse_filename(const char *filename, QDict *options,
     int fat_type = 0;
     bool floppy = false;
     bool rw = false;
-    bool no_mbr = false;
     int i;
 
     if (!strstart(filename, "fat:", NULL)) {
@@ -1138,10 +1126,6 @@ static void vvfat_parse_filename(const char *filename, QDict *options,
         rw = true;
     }
 
-    if (strstr(filename, ":no-mbr:")) {
-        no_mbr = true;
-    }
-
     /* Get the directory name without options */
     i = strrchr(filename, ':') - filename;
     assert(i >= 3);
@@ -1157,7 +1141,6 @@ static void vvfat_parse_filename(const char *filename, QDict *options,
     qdict_put_int(options, "fat-type", fat_type);
     qdict_put_bool(options, "floppy", floppy);
     qdict_put_bool(options, "rw", rw);
-    qdict_put_bool(options, "no-mbr", no_mbr);
 }
 
 static void vvfat_get_size_paremeters(uint64_t size, BDRVVVFATState *s, bool floppy,
@@ -1338,7 +1321,7 @@ static int vvfat_open(BlockDriverState *bs, QDict *options, int flags,
 
     vvfat_get_size_paremeters(size, s, floppy, errp);
 
-    if (!floppy && !qemu_opt_get_bool(opts, "no-mbr", false)) {
+    if (!floppy) {
         /* We have an MBR in this case, it stay zero otherwise.. */
         s->offset_to_bootsector = 0x3f;
     }
@@ -1546,14 +1529,6 @@ read_cluster_directory:
             s->current_cluster = -1;
             return -1;
         }
-        if (cluster_num == s->current_mapping->end - 1
-            && s->current_mapping->info.file.last_cluster_data) {
-            memcpy (s->cluster + result,
-                    s->current_mapping->info.file.last_cluster_data + result,
-                    s->cluster_size - result);
-        } else {
-            memset (s->cluster + result, 0, s->cluster_size - result);
-        }
         s->current_cluster = cluster_num;
     }
     return 0;
@@ -1734,11 +1709,6 @@ typedef struct commit_t {
         ACTION_RENAME, ACTION_WRITEOUT, ACTION_NEW_FILE, ACTION_MKDIR
     } action;
 } commit_t;
-
-#ifdef DEBUG
-static const char * const action_str[] =
-  { "rename", "writeout", "new_file", "mkdir" };
-#endif
 
 static void clear_commits(BDRVVVFATState* s)
 {
@@ -2033,7 +2003,6 @@ get_cluster_count_for_direntry(BDRVVVFATState* s, direntry_t* direntry, const ch
      */
     int copy_it = 0;
     int was_modified = 0;
-    int size_was_modified = 0;
     int32_t ret = 0;
 
     uint32_t cluster_num = begin_of_direntry(direntry);
@@ -2091,11 +2060,6 @@ get_cluster_count_for_direntry(BDRVVVFATState* s, direntry_t* direntry, const ch
 
             assert(mapping->mode & MODE_NORMAL);
 
-            d = (direntry_t*)array_get(&(s->directory), mapping->dir_index);
-            if ((filesize_of_direntry(d) & 0x1ff)
-                != (filesize_of_direntry(direntry) & 0x1ff)) {
-                size_was_modified = 1;
-            }
         } else if (is_file(direntry))
             /* new file */
             schedule_new_file(s, g_strdup(path), cluster_num);
@@ -2182,15 +2146,10 @@ get_cluster_count_for_direntry(BDRVVVFATState* s, direntry_t* direntry, const ch
 
         cluster_num = modified_fat_get(s, cluster_num);
 
-        if (fat_eof(s, cluster_num)) {
-            if (size_was_modified && !was_modified) {
-                schedule_writeout(s, mapping->dir_index,
-                                  filesize_of_direntry(direntry) & ~0x1ff);
-            }
+        if (fat_eof(s, cluster_num))
             return ret;
-        } else if (cluster_num < 2 || cluster_num > s->max_fat_value - 16) {
+        else if (cluster_num < 2 || cluster_num > s->max_fat_value - 16)
             return -1;
-        }
 
         offset += s->cluster_size;
     }
@@ -2469,12 +2428,8 @@ static int remove_mapping(BDRVVVFATState* s, int mapping_index)
     mapping_t* first_mapping = array_get(&(s->mapping), 0);
 
     /* free mapping */
-    if (mapping->first_mapping_index < 0)
-        free(mapping->path);
-    if ((mapping->mode & MODE_NORMAL)
-        && mapping->info.file.last_cluster_data) {
-        free (mapping->info.file.last_cluster_data);
-        mapping->info.file.last_cluster_data = NULL;
+    if (mapping->first_mapping_index < 0) {
+        g_free(mapping->path);
     }
 
     /* remove from s->mapping */
@@ -2596,11 +2551,10 @@ static int commit_mappings(BDRVVVFATState* s,
                         mapping->info.dir.first_dir_index +
                         0x10 * s->sectors_per_cluster *
                         (mapping->end - mapping->begin);
-            } else {
+            } else
                 next_mapping->info.file.offset = mapping->info.file.offset +
                         (mapping->end - mapping->begin);
-                next_mapping->info.file.last_cluster_data = NULL;
-            }
+
             mapping = next_mapping;
         }
 
@@ -2661,11 +2615,17 @@ commit_direntries(BDRVVVFATState* s, int dir_index, int parent_mapping_index)
                 factor * (old_cluster_count - new_cluster_count));
 
     for (c = first_cluster; !fat_eof(s, c); c = modified_fat_get(s, c)) {
+        direntry_t *first_direntry;
+
         direntry = array_get(&(s->directory), current_dir_index);
         ret = vvfat_read(s->bs, cluster2sector(s, c), (uint8_t *)direntry,
                 s->sectors_per_cluster);
         if (ret)
             return ret;
+
+        /* The first directory entry on the filesystem is the volume name */
+        first_direntry = (direntry_t*) s->directory.pointer;
+        assert(!memcmp(first_direntry->name, s->volume_label, 11));
 
         current_dir_index += factor;
     }
@@ -2714,9 +2674,6 @@ commit_one_file(BDRVVVFATState* s, int dir_index, uint32_t offset)
         return -1;
     }
 
-    DLOG(fprintf(stderr,"commit_one_file: %s at %d (len=%d)\n",
-                 mapping->path, offset, size));
-
     for (i = 0; i < offset; i += s->cluster_size) {
         c = modified_fat_get(s, c);
     }
@@ -2762,15 +2719,6 @@ commit_one_file(BDRVVVFATState* s, int dir_index, uint32_t offset)
             return -2;
         }
 
-        if (rest_size != s->cluster_size) {
-            mapping = find_mapping_for_cluster(s, c);
-            assert (mapping);
-            if (!mapping->info.file.last_cluster_data)
-                mapping->info.file.last_cluster_data=malloc (s->cluster_size);
-            memcpy (mapping->info.file.last_cluster_data, cluster,
-                    s->cluster_size);
-        }
-
         offset += rest_size;
         c = c1;
     }
@@ -2795,7 +2743,7 @@ static void check1(BDRVVVFATState* s)
     for (i = 0; i < s->mapping.next; i++) {
         mapping_t* mapping = array_get(&(s->mapping), i);
         if (mapping->mode & MODE_DELETED) {
-            fprintf(stderr, "deleted: %s\n", mapping->path);
+            fprintf(stderr, "deleted\n");
             continue;
         }
         assert(mapping->dir_index < s->directory.next);
@@ -2864,9 +2812,9 @@ static int handle_renames_and_mkdirs(BDRVVVFATState* s)
     fprintf(stderr, "handle_renames\n");
     for (i = 0; i < s->commits.next; i++) {
         commit_t* commit = array_get(&(s->commits), i);
-        fprintf(stderr, "%d, %s (%u, %s)\n", i,
+        fprintf(stderr, "%d, %s (%u, %d)\n", i,
                 commit->path ? commit->path : "(null)",
-                commit->param.rename.cluster, action_str[commit->action]);
+                commit->param.rename.cluster, commit->action);
     }
 #endif
 
@@ -3035,7 +2983,6 @@ static int coroutine_fn GRAPH_RDLOCK handle_commits(BDRVVVFATState* s)
             mapping->read_only = 0;
             mapping->mode = MODE_NORMAL;
             mapping->info.file.offset = 0;
-            mapping->info.file.last_cluster_data = NULL;
 
             if (commit_one_file(s, j, 0)) {
                 fail = -7;
@@ -3065,20 +3012,21 @@ static int handle_deletes(BDRVVVFATState* s)
         for (i = 1; i < s->mapping.next; i++) {
             mapping_t* mapping = array_get(&(s->mapping), i);
             if (mapping->mode & MODE_DELETED) {
-                DLOG(direntry_t* entry = array_get(&(s->directory),
-                                                   mapping->dir_index));
-                /* remove file/directory */
-                if (mapping->mode & MODE_DIRECTORY) {
-                    int j, next_dir_index = s->directory.next,
+                direntry_t* entry = array_get(&(s->directory),
+                        mapping->dir_index);
+
+                if (is_free(entry)) {
+                    /* remove file/directory */
+                    if (mapping->mode & MODE_DIRECTORY) {
+                        int j, next_dir_index = s->directory.next,
                         first_dir_index = mapping->info.dir.first_dir_index;
 
                         if (rmdir(mapping->path) < 0) {
                             if (errno == ENOTEMPTY) {
                                 deferred++;
                                 continue;
-                            } else {
+                            } else
                                 return -5;
-                            }
                         }
 
                         for (j = 1; j < s->mapping.next; j++) {
@@ -3095,6 +3043,7 @@ static int handle_deletes(BDRVVVFATState* s)
                                 next_dir_index - first_dir_index);
 
                         deleted++;
+                    }
                 } else {
                     if (unlink(mapping->path))
                         return -4;
@@ -3120,6 +3069,10 @@ static int handle_deletes(BDRVVVFATState* s)
 static int coroutine_fn GRAPH_RDLOCK do_commit(BDRVVVFATState* s)
 {
     int ret = 0;
+
+    /* the real meat are the commits. Nothing to do? Move along! */
+    if (s->commits.next == 0)
+        return 0;
 
     vvfat_close_current_file(s);
 
@@ -3167,10 +3120,8 @@ static int coroutine_fn GRAPH_RDLOCK try_commit(BDRVVVFATState* s)
 {
     vvfat_close_current_file(s);
 DLOG(checkpoint());
-    if(!is_consistent(s)) {
-        DLOG(fprintf(stderr,"not consistent\n"));
+    if(!is_consistent(s))
         return -1;
-    }
     return do_commit(s);
 }
 
@@ -3430,15 +3381,8 @@ static void vvfat_child_perm(BlockDriverState *bs, BdrvChild *c,
 static void vvfat_close(BlockDriverState *bs)
 {
     BDRVVVFATState *s = bs->opaque;
-    int i;
 
     vvfat_close_current_file(s);
-
-    for (i = 0; i < s->mapping.next; i++) {
-	mapping_t *mapping=array_get(&(s->mapping),i);
-	if (mapping->first_mapping_index < 0)
-	    free (mapping->path);
-    }
     array_free(&(s->fat));
     array_free(&(s->directory));
     array_free(&(s->mapping));
